app/requirements.txt
# Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.9

# Redis
redis==5.1.1
aioredis==2.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Agent / LLM
langchain==0.3.1
langchain-anthropic==0.2.1
langchain-community==0.3.1
langgraph==0.2.28

# Tools
tavily-python==0.5.0
pandas==2.2.3
openpyxl==3.1.5

# Voice / Whisper
openai-whisper==20240930
soundfile==0.12.1

# Utils
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
websockets==13.1
aiofiles==24.1.0